+++ js/main.js
@@ -3,7 +3,7 @@
 import { randomNumber, randomFloatNumber } from './util.js';
 import { fillingArrObj } from './data.js';
 import { renderCard } from './renderCard.js';
-import { mapsChanges } from './changeForm.js'; 
+import { mapsChanges } from './map.js'; 
 
 console.log('Рандомное число ', randomNumber(20, 100));
 console.log('Рандомное число с плавающей точкой ', randomFloatNumber(1.3, 1.67, 5));
